% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_helpers.R
\name{add_provincial_abbreviations}
\alias{add_provincial_abbreviations}
\title{Add provincial abbreviations as factor}
\usage{
add_provincial_abbreviations(data)
}
\arguments{
\item{data}{code{cansim} package data frame with provincial level data}
}
\value{
a code{cansim} package data frame with additional factor GEO.abb that contains language-specific provincial abbreviations
}
\description{
Add provincial abbreviations as factor
}
