% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_helpers.R
\name{add_provincial_appreviations}
\alias{add_provincial_appreviations}
\title{add provincial appreviation as factor}
\usage{
add_provincial_appreviations(data)
}
\arguments{
\item{data}{cansim data frame with provincial level data}
}
\value{
a cansim data frame with additional factor GEO.abb that contains language-specific provincial appreviations
}
\description{
add provincial appreviation as factor
}
