% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{adjust_cansim_values_by_variable}
\alias{adjust_cansim_values_by_variable}
\title{Adjust CANSIM Value by scaled amount
French part does not work, probably encoding issues
Legacy function}
\usage{
adjust_cansim_values_by_variable(data, var)
}
\description{
Adjust CANSIM Value by scaled amount
French part does not work, probably encoding issues
Legacy function
}
