% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{adjust_cansim_values_by_variable_old}
\alias{adjust_cansim_values_by_variable_old}
\title{Adjust Cansim Value by scaled amount
French part does not work, probably encoding issues}
\usage{
adjust_cansim_values_by_variable_old(data, var)
}
\description{
Adjust Cansim Value by scaled amount
French part does not work, probably encoding issues
}
