% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_helpers.R
\name{cache_date_format}
\alias{cache_date_format}
\title{Cache date format for a table}
\usage{
cache_date_format(table_number, format_type)
}
\arguments{
\item{table_number}{the table number}

\item{format_type}{the detected format type}
}
\description{
Cache date format for a table
}
\keyword{internal}
