% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{cansim_old_to_new}
\alias{cansim_old_to_new}
\title{Translate deprecated CANSIM table number into new NDM-format table catalogue number}
\usage{
cansim_old_to_new(oldCansimTableNumber)
}
\arguments{
\item{oldCansimTableNumber}{deprecated style CANSIM table number (e.g. "427-0001")}
}
\value{
A character string with the new-format NDM table number
}
\description{
Returns NDM table catalogue equivalent given a standard old-format CANSIM table number
}
\examples{
cansim_old_to_new("026-0018")
}
