% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{cansim_old_to_new2}
\alias{cansim_old_to_new2}
\title{translate from old table number to NDM table number by web search and scraping}
\usage{
cansim_old_to_new2(oldCansimTableNumber)
}
\arguments{
\item{oldCansimTableNumber}{the old cansim table number}
}
\description{
translate from old table number to NDM table number by web search and scraping
}
