% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\docType{data}
\name{correspondence}
\alias{correspondence}
\title{The correspondence file for old to new StatCan table numbers is included in the package}
\description{
The correspondence file for old to new StatCan table numbers is included in the package
}
\references{
\url{https://www.statcan.gc.ca/eng/developers-developpeurs/cansim_id-product_id-concordance.csv}
}
\author{
Statistics Canada
}
\keyword{data}
\keyword{internal}
