% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_sql.R
\name{create_index}
\alias{create_index}
\title{create database index}
\usage{
create_index(connection, table_name, field)
}
\arguments{
\item{connection}{connection to database}

\item{table_name}{sql table name}

\item{field}{name of field to index}
}
\value{
`NULL``
}
\description{
create database index
}
\keyword{internal}
