% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_sql.R
\name{disconnect_cansim_sqlite}
\alias{disconnect_cansim_sqlite}
\title{Disconnect from a cansim database connection (deprecated)}
\usage{
disconnect_cansim_sqlite(connection)
}
\arguments{
\item{connection}{connection to database}
}
\value{
`NULL``
}
\description{
This method is deprecated and will be removed in a future version, please use
`disconnect_cansim_connection()` instead.
}
\examples{
\dontrun{
con <- get_cansim_connection("34-10-0013", format="sqlite")
disconnect_cansim_connection(con)
}
}
\keyword{internal}
