% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{fold_in_metadata_for_columns}
\alias{fold_in_metadata_for_columns}
\title{Fold in metadata and for selected columns}
\usage{
fold_in_metadata_for_columns(data, data_path, column_names)
}
\arguments{
\item{data}{the data table}

\item{data_path}{base path to save parsed metadata}

\item{column_names}{the names of the columns}
}
\value{
data table including the metadata information
}
\description{
Fold in metadata and for selected columns
}
\keyword{internal}
