% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_helpers.R
\name{get_cached_connection_metadata}
\alias{get_cached_connection_metadata}
\title{Get cached connection metadata}
\usage{
get_cached_connection_metadata(cache_key)
}
\arguments{
\item{cache_key}{unique key for this connection}
}
\value{
cached metadata or NULL
}
\description{
Get cached connection metadata
}
\keyword{internal}
