% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_helpers.R
\name{get_cached_date_format}
\alias{get_cached_date_format}
\title{Get cached date format for a table}
\usage{
get_cached_date_format(table_number)
}
\arguments{
\item{table_number}{the table number}
}
\value{
date format string or NULL
}
\description{
Get cached date format for a table
}
\keyword{internal}
