% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_tables_list.R
\name{get_cansim_key_release_schedule}
\alias{get_cansim_key_release_schedule}
\title{Major economic indicator release schedule}
\usage{
get_cansim_key_release_schedule()
}
\value{
a tibble with data, and details for major economic indicator release

Returns \code{NULL} if the data could not be retrieved because StatCan is unavailable.
}
\description{
Returns every release date of major economic indicators since March 14, 2012.
This also includes scheduled future releases.
}
\examples{
\donttest{
get_cansim_key_release_schedule()
}

}
