% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_ndm}
\alias{get_cansim_ndm}
\title{Retrieve a CANSIM table using an NDM number}
\usage{
get_cansim_ndm(cansimTableNumber, language = "english",
  refresh = FALSE)
}
\arguments{
\item{cansimTableNumber}{the NDM table number to load}

\item{language}{\code{"en"} or \code{"english"} for English and \code{"fr"} or \code{"french"} for French language versions (defaults to English)}

\item{refresh}{Optionally force reload of CANSIM data, default is \code{FALSE}}
}
\value{
a tibble with retrieved NDM table data
}
\description{
Retrieves a CANSIM table as a tidy dataframe using an NDM number. Retrieved table data is cached for the duration of the current R session only by default.
}
\examples{
get_cansim("34-10-0013")

}
