% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_old}
\alias{get_cansim_old}
\title{Legacy method, get cansim tables based on old table names
Get cansim table into tidy dataframe
Caches the table data for the current session}
\usage{
get_cansim_old(cansimTableNumber, language = "english")
}
\description{
Legacy method, get cansim tables based on old table names
Get cansim table into tidy dataframe
Caches the table data for the current session
}
