% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_table_last_release_date}
\alias{get_cansim_table_last_release_date}
\title{Get the latest release data for a StatCan table, if available}
\usage{
get_cansim_table_last_release_date(cansimTableNumber)
}
\arguments{
\item{cansimTableNumber}{the NDM table number}
}
\value{
A datetime object if a release data is available, NULL otherwise.

Returns \code{NULL} if the data could not be retrieved because StatCan is unavailable.
}
\description{
This can be used to check when a table has last been updated.
}
\examples{
\donttest{
get_cansim_table_last_release_date("34-10-0013")
}
}
