% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_tables_list.R
\name{get_cansim_table_list}
\alias{get_cansim_table_list}
\title{Retrieve list of all Statistics Canada data tables}
\usage{
get_cansim_table_list()
}
\description{
Internal function to collect an up-to-date and complete list of Statistics Canada data tables
}
