% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_tables_list.R
\name{get_cansim_table_list_page}
\alias{get_cansim_table_list_page}
\title{get page of cansim tables}
\usage{
get_cansim_table_list_page(start = 0, rows = 1000)
}
\description{
get page of cansim tables
}
