% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_vectors.R
\name{get_cansim_vector_info}
\alias{get_cansim_vector_info}
\title{Retrieve metadata for specified Statistics Canada data vectors}
\usage{
get_cansim_vector_info(vectors)
}
\arguments{
\item{vectors}{a vector of cansim vectors}
}
\value{
A tibble with metadata for selected vectors

Returns \code{NULL} if the data could not be retrieved because StatCan is unavailable.
}
\description{
Allows for the retrieval of metadata for Statistics Canada data vectors
}
\examples{
\donttest{
get_cansim_vector_info("v41690973")
}
}
