% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_helpers.R
\name{get_deduped_column_level_data}
\alias{get_deduped_column_level_data}
\title{Get column names de-duplicated and in the correct order}
\usage{
get_deduped_column_level_data(cansimTableNumber, language, column)
}
\arguments{
\item{cansimTableNumber}{The table number}

\item{language}{The language of the column names}

\item{column}{The column name}
}
\value{
A tibble with the column names
}
\description{
Get column names de-duplicated and in the correct order
}
\keyword{internal}
