% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_parquet.R
\name{list_cansim_cached_tables}
\alias{list_cansim_cached_tables}
\title{List cached cansim arrow and SQLite databases}
\usage{
list_cansim_cached_tables(
  cache_path = Sys.getenv("CANSIM_CACHE_PATH"),
  refresh = FALSE
)
}
\arguments{
\item{cache_path}{Optional, default value is `Sys.getenv('CANSIM_CACHE_PATH')`.}

\item{refresh}{Optional, refresh the last updated date of cached cansim tables}
}
\value{
A tibble with the list of all tables that are currently cached at the given cache path.
}
\description{
List cached cansim arrow and SQLite databases
}
\examples{
\dontrun{
list_cansim_cached_tables()
}
}
