% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_parquet.R
\name{list_cansim_dbs}
\alias{list_cansim_dbs}
\title{List cached cansim arrow and SQlite databases}
\usage{
list_cansim_dbs(cache_path = getOption("cansim.cache_path"), refresh = FALSE)
}
\arguments{
\item{cache_path}{Optional, default value is `getOption("cansim.cache_path")`.}

\item{refresh}{Optional, refresh the last updated date of cached cansim tables}
}
\value{
A tibble with the list of all tables that are currently cached at the given cache path.
}
\description{
List cached cansim arrow and SQlite databases
}
\examples{
\dontrun{
list_cansim_dbs()
}
}
