% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{normailze_cansim_values}
\alias{normailze_cansim_values}
\title{normailzes cancsim values by setting all units to counts/dollars instead of millions, etc.
if "replace" is true, it will replace the VALUE field with normailzed values and drop the scale columns,
otherwise it keeps the scale columns and creased a new column names "NORMALIZED_VALUE" with the normalized value}
\usage{
normailze_cansim_values(data, replace = TRUE)
}
\description{
normailzes cancsim values by setting all units to counts/dollars instead of millions, etc.
if "replace" is true, it will replace the VALUE field with normailzed values and drop the scale columns,
otherwise it keeps the scale columns and creased a new column names "NORMALIZED_VALUE" with the normalized value
}
