% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{parse_and_fold_in_metadata}
\alias{parse_and_fold_in_metadata}
\title{Parse metadata and fold into data table}
\usage{
parse_and_fold_in_metadata(data, meta, data_path)
}
\arguments{
\item{data}{the data table}

\item{meta}{the raw metadata table}

\item{data_path}{base path to save parsed metadata}
}
\value{
data table including the metadata information
}
\description{
Parse metadata and fold into data table
}
\keyword{internal}
