% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{parse_metadata}
\alias{parse_metadata}
\title{Parse metadata}
\usage{
parse_metadata(meta, data_path)
}
\arguments{
\item{meta}{the raw metadata table}

\item{data_path}{base path to save parsed metadata}
}
\description{
Parse metadata
}
\keyword{internal}
