% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_helpers.R
\name{set_cached_connection_metadata}
\alias{set_cached_connection_metadata}
\title{Set cached connection metadata}
\usage{
set_cached_connection_metadata(cache_key, metadata)
}
\arguments{
\item{cache_key}{unique key for this connection}

\item{metadata}{metadata to cache}
}
\description{
Set cached connection metadata
}
\keyword{internal}
