% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{adjust_cansim_values_by_variable}
\alias{adjust_cansim_values_by_variable}
\title{(Legacy function) Adjust CANSIM table values by scaled amount; however
French does not work, probably due to encoding issues.}
\usage{
adjust_cansim_values_by_variable(data, var)
}
\arguments{
\item{data}{???}

\item{var}{???}
}
\description{
(Legacy function) Adjust CANSIM table values by scaled amount; however
French does not work, probably due to encoding issues.
}
