% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{cansim_old_to_new}
\alias{cansim_old_to_new}
\title{translate from old table number to NDM table number}
\usage{
cansim_old_to_new(oldCansimTableNumber)
}
\description{
translate from old table number to NDM table number
}
