% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{cansim_old_to_new}
\alias{cansim_old_to_new}
\title{translate from old table number to NDM table number}
\usage{
cansim_old_to_new(oldCansimTableNumber)
}
\arguments{
\item{oldCansimTableNumber}{the old cansim table number. Returns the corresponding NDM number}
}
\value{
cansim ndm table number
}
\description{
translate from old table number to NDM table number
}
