% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{cansim_old_to_new}
\alias{cansim_old_to_new}
\title{Translate CANSIM table number into NDM table}
\usage{
cansim_old_to_new(oldCansimTableNumber)
}
\arguments{
\item{oldCansimTableNumber}{the old CANSIM table number. Returns the corresponding NDM number.}
}
\value{
CANSIM NDM table number
}
\description{
Returns NDM table equivalent given a standard old-format CANSIM table number
}
