% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_parquet.R
\name{csv2arrow}
\alias{csv2arrow}
\title{convert csv to arrow}
\usage{
csv2arrow(
  csv_file,
  arrow_file,
  format = "parquet",
  col_names,
  value_column = "VALUE",
  na = c(NA, "..", "", "...", "F"),
  text_encoding = "UTF-8",
  delim = ","
)
}
\arguments{
\item{csv_file}{input csv path}

\item{arrow_file}{output arrow database path}

\item{format}{format of arrow file, "parquet" or "feather" (default parquet)}

\item{col_names}{column names of the csv file}

\item{value_column}{name of the value column with numeric data}

\item{na}{na character strings}

\item{text_encoding}{encoding of csv file (default UTF-8)}

\item{delim}{(Optional) csv deliminator, default is ","}
}
\value{
A database connection
}
\description{
convert csv to arrow
}
\keyword{internal}
