% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_sql.R
\name{disconnect_cansim_sqlite}
\alias{disconnect_cansim_sqlite}
\title{Disconnect from a cansim database connection}
\usage{
disconnect_cansim_sqlite(connection)
}
\arguments{
\item{connection}{connection to database}
}
\value{
`NULL``
}
\description{
Disconnect from a cansim database connection
}
\examples{
\dontrun{
con <- get_cansim_sqlite("34-10-0013")
disconnect_cansim_sqlite(con)
}
}
