% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{fold_in_metadata_for_columns}
\alias{fold_in_metadata_for_columns}
\title{Fold in metadata and for selected columns}
\usage{
fold_in_metadata_for_columns(data, data_path, column_names)
}
\arguments{
\item{data}{A tibble with StatCan table data as e.g. returned by \code{get_cansim}.}

\item{data_path}{base path to save parsed metadata}

\item{column_names}{the names of the columns}
}
\value{
A tibble including the metadata information
}
\description{
Fold in metadata and for selected columns
}
\keyword{internal}
