% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim}
\alias{get_cansim}
\title{Get cansim table into tidy dataframe}
\usage{
get_cansim(cansimTableNumber, language = "english", refresh = FALSE)
}
\arguments{
\item{cansimTableNumber}{the table number to load, accepts old or new NDM table numbers}

\item{language}{\code{"en"} or \code{"english"} for English and \code{"fr"} or \code{"french"} for French language versions. Defaults to English.}

\item{refresh}{Optionally force reload of CANSIM data, default is \code{FALSE}. CANSIM data is cached for the duration of the R session only}
}
\value{
a tibble with the cansim table data
}
\description{
Caches the table data for the current session
}
