% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_changed_tables}
\alias{get_cansim_changed_tables}
\title{get list of changed tables in date range}
\usage{
get_cansim_changed_tables(start_date)
}
\arguments{
\item{start_date}{starting date to look for changes that changed on or after that date}
}
\value{
a tible with cansim product ids and release times
}
\description{
get list of changed tables in date range
}
