% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_helpers.R
\name{get_cansim_code_set}
\alias{get_cansim_code_set}
\title{Get NDM code sets}
\usage{
get_cansim_code_set(
  code_set = c("scalar", "frequency", "symbol", "status", "uom", "survey", "subject",
    "wdsResponseStatus"),
  refresh = FALSE
)
}
\arguments{
\item{code_set}{the code set to retrieve.}

\item{refresh}{Default is \code{FALSE}, repeated calls during the same session will hit the cached data.
To refresh the code list during a running R session set to \code{TRUE}}
}
\value{
A tibble with english and french labels for the given code set
}
\description{
Useful to get a list of surveys  or subjects and used internally
}
\examples{
\donttest{
get_cansim_code_set("survey")
}
}
