% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_cube_metadata}
\alias{get_cansim_cube_metadata}
\title{Get table metadata from API}
\usage{
get_cansim_cube_metadata(cansimTableNumber)
}
\arguments{
\item{cansimTableNumber}{new or old CANSIM table number}
}
\value{
a tibble containing the table metadata
}
\description{
Get table metadata from API
}
