% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_cube_metadata}
\alias{get_cansim_cube_metadata}
\title{Get table metadata from API
Also accepts a vector of cansim table numbers, new or old.
Patience required, StatCan API is very slow. Alternatively
one can also use the `list_cansim_tables()` function to retieve
a (cached) list of cansim tables with (more limited) metadata}
\usage{
get_cansim_cube_metadata(cansimTableNumber)
}
\arguments{
\item{cansimTableNumber}{new or old CANSIM table number}
}
\value{
a tibble containing the table metadata
}
\description{
Get table metadata from API
Also accepts a vector of cansim table numbers, new or old.
Patience required, StatCan API is very slow. Alternatively
one can also use the `list_cansim_tables()` function to retieve
a (cached) list of cansim tables with (more limited) metadata
}
