% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_data_for_table_coord_periods}
\alias{get_cansim_data_for_table_coord_periods}
\title{get data for cansim table number and coordinate for latest periods}
\usage{
get_cansim_data_for_table_coord_periods(cansimTableNumber, coordinate,
  periods = 1)
}
\arguments{
\item{cansimTableNumber}{cansim table number}

\item{coordinate}{cansim coordinate}

\item{periods}{number of latest periods to retrieve data for}
}
\value{
a tibble with data
}
\description{
get data for cansim table number and coordinate for latest periods
}
