% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_metadata.R
\name{get_cansim_series_info_cube_coord}
\alias{get_cansim_series_info_cube_coord}
\title{Retrieve series info for given table id and coordinates}
\usage{
get_cansim_series_info_cube_coord(
  cansimTableNumber,
  coordinates,
  timeout = 1000,
  refresh = FALSE
)
}
\arguments{
\item{cansimTableNumber}{A new or old CANSIM/NDM table number or a vector of table numbers}

\item{coordinates}{A vector of coordinates}

\item{timeout}{Timeout for the API call}

\item{refresh}{Refresh the data from the Statistics Canada API}
}
\value{
a tibble containing the table template
}
\description{
Retrieves series information by coordinates
}
\examples{
\dontrun{
get_cansim_table_template("34-10-0013")
}
}
