% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_table_last_release_date}
\alias{get_cansim_table_last_release_date}
\title{Get the latest release data for a StatCan table, if available}
\usage{
get_cansim_table_last_release_date(cansimTableNumber)
}
\arguments{
\item{cansimTableNumber}{the NDM table number}
}
\value{
A datatime object if a release data is available, NULL otherwise.
}
\description{
This can be used to check when a table has last been updated.
}
\examples{
\dontrun{
get_cansim_table_last_release_date("34-10-0013")
}
}
