% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_tables_list.R
\name{get_cansim_table_list_page}
\alias{get_cansim_table_list_page}
\title{internal method to get page of cansim tables}
\usage{
get_cansim_table_list_page(start_offset = 0, max_rows = 1000)
}
\arguments{
\item{start_offset}{starting row}

\item{max_rows}{number of rows, maximum is 1000}
}
\description{
internal method to get page of cansim tables
}
