% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_table_notes}
\alias{get_cansim_table_notes}
\title{Get CANSIM table notes}
\usage{
get_cansim_table_notes(cansimTableNumber, language = "english",
  refresh = FALSE)
}
\arguments{
\item{cansimTableNumber}{the NDM table number to load}

\item{language}{\code{"en"} or \code{"english"} for English and \code{"fr"} or \code{"french"} for French language versions. Defaults to English.}

\item{refresh}{Optionally force reload of CANSIM data, default is \code{FALSE}. CANSIM data is cached for the duration of the R session only}
}
\description{
Get CANSIM table notes
}
