% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_table_overview}
\alias{get_cansim_table_overview}
\title{Get CANSIM table overview text}
\usage{
get_cansim_table_overview(cansimTableNumber, language = "english",
  refresh = FALSE)
}
\arguments{
\item{cansimTableNumber}{the NDM table number to load}

\item{language}{\code{"en"} or \code{"english"} for English and \code{"fr"} or \code{"french"} for French language versions. Defaults to English.}

\item{refresh}{Optionally force reload of CANSIM data, default is \code{FALSE}. CANSIM data is cached for the duration of the R session only}
}
\value{
none
}
\description{
Needs to load the whole CANSIM table in order to display overview information. Prints overview as message
}
