% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_table_overview}
\alias{get_cansim_table_overview}
\title{Get cansim table overview text}
\usage{
get_cansim_table_overview(cansimTableNumber, column, language = "english",
  refresh = FALSE)
}
\arguments{
\item{cansimTableNumber}{the NDM table number to load}

\item{language}{"en" or "fr" for english or french language version. Defaults to english.}

\item{refresh}{Optionally force reload of cansim data, default is *FALSE*. Cansim data is cached for the duration of the R session only}
}
\value{
none
}
\description{
Needs to load the whole cansim table in order to dusplay overview information. Prints overview as message
}
