% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_table_subject}
\alias{get_cansim_table_subject}
\title{Retrieve Statistics Canada data table subject detail}
\usage{
get_cansim_table_subject(cansimTableNumber, language = "english",
  refresh = FALSE)
}
\arguments{
\item{cansimTableNumber}{the NDM table number to load}

\item{language}{\code{"en"} or \code{"english"} for English and \code{"fr"} or \code{"french"} for French language versions (default set to English)}

\item{refresh}{(Optional) When set to \code{TRUE}, forces a reload of data table (default is \code{FALSE})}
}
\description{
Returns table subject detail given an NDM table number in English or French. Retrieved table information data is cached for the duration of the R session only.
}
\examples{
get_cansim_table_subject("34-10-0013")

}
