% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_table_survey}
\alias{get_cansim_table_survey}
\title{Get cansim table survey}
\usage{
get_cansim_table_survey(cansimTableNumber, language = "english",
  refresh = FALSE)
}
\arguments{
\item{cansimTableNumber}{the NDM table number to load}

\item{language}{"en" or "fr" for english or french language version. Defaults to english.}

\item{refresh}{Optionally force reload of cansim data, default is *FALSE*. Cansim data is cached for the duration of the R session only}
}
\description{
Get cansim table survey
}
