% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_table_url}
\alias{get_cansim_table_url}
\title{get table url from API, should be more stable than guessing the link}
\usage{
get_cansim_table_url(cansimTableNumber, language)
}
\description{
get table url from API, should be more stable than guessing the link
}
