% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_vector}
\alias{get_cansim_vector}
\title{Retrieve data for a CANSIM vector released within a given time frame}
\usage{
get_cansim_vector(vectors, start_time, end_time = Sys.Date())
}
\arguments{
\item{vectors}{The list of vectors to retrieve}

\item{start_time}{Starting date in \code{YYYY-MM-DD} format to look for changes that changed on or after that date}

\item{end_time}{Optional ending time in \code{YYYY-MM-DD} format (defaults to current system time)}
}
\value{
A tibble with data for vectors released between start and end time
}
\description{
Retrieve data for a CANSIM vector released within a given time frame
}
