% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{get_cansim_vector_for_latest_periods}
\alias{get_cansim_vector_for_latest_periods}
\title{get data for cansim vector(s) for last N periods}
\usage{
get_cansim_vector_for_latest_periods(vectors, periods = 1)
}
\arguments{
\item{vectors}{list of vectors to retrieve}

\item{periods}{number of latest periods to retrieve data for}
}
\value{
a tibble with data for vector(s) for last N periods
}
\description{
get data for cansim vector(s) for last N periods
}
