% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_vectors.R
\name{get_cansim_vector_for_latest_periods}
\alias{get_cansim_vector_for_latest_periods}
\title{Retrieve data for specified Statistics Canada data vector(s) for last N periods}
\usage{
get_cansim_vector_for_latest_periods(vectors, periods = 1)
}
\arguments{
\item{vectors}{The list of vectors to retrieve}

\item{periods}{Numeric value for number of latest periods to retrieve data for}
}
\value{
A tibble with data for specified vector(s) for the last N periods
}
\description{
Allows for the retrieval of data for specified vector series for the N most-recently released periods.
}
\examples{
get_cansim_vector_for_latest_periods("v41690973",10)

}
