% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_vectors.R
\name{get_cansim_vector_info}
\alias{get_cansim_vector_info}
\title{Retrieve metadatadata for specified CANSIM vectors}
\usage{
get_cansim_vector_info(vectors)
}
\arguments{
\item{vectors}{a vector of cansim vectors}
}
\value{
A tibble with metadata for the cansim vectors
}
\description{
Allows for the retrieval of metadatadata for CANSIM vectors
}
