% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim_sql.R
\name{list_cansim_sqlite_cached_tables}
\alias{list_cansim_sqlite_cached_tables}
\title{List cached cansim SQLite database}
\usage{
list_cansim_sqlite_cached_tables(cache_path = getOption("cansim.cache_path"))
}
\arguments{
\item{cache_path}{Optional, default value is `getOption("cansim.cache_path")`.}
}
\value{
A tibble with the list of all tables that are currently cached at the given cache path.
}
\description{
List cached cansim SQLite database
}
\examples{
\dontrun{
list_cansim_sqlite_cached_tables()
}
}
