% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{view_cansim_webpage}
\alias{view_cansim_webpage}
\title{View CANSIM table information in browser}
\usage{
view_cansim_webpage(cansimTableNumber)
}
\arguments{
\item{cansimTableNumber}{CANSIM or NDM table number}
}
\value{
none
}
\description{
Opens CANSIM table on Statistics Canada's website using default browser. This may be useful for getting further info on CANSIM table and survey methods.
}
\examples{
\donttest{
view_cansim_webpage("34-10-0013")
}
}
