% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cansim.R
\name{view_cansim_webpage}
\alias{view_cansim_webpage}
\title{open cansim table information in browser}
\usage{
view_cansim_webpage(cansimTableNumber, browser = getOption("browser"))
}
\arguments{
\item{cansimTableNumber}{cansim table number}

\item{browser}{optionally a browser to open the page in}
}
\value{
none
}
\description{
useful for getting furthe info on cansim table and survey methods
}
